chore: tidy tests
